wrap login with titleprovider component
